% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{ASSIGN}
\alias{ASSIGN}
\title{Atribui e retorna um valor}
\usage{
ASSIGN(x)
}
\arguments{
\item{x}{Valor de entrada.}
}
\value{
O mesmo valor \code{x}.
}
\description{
Atribui e retorna um valor
}
