% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{AUTONUMBER}
\alias{AUTONUMBER}
\title{Gera sequência numérica automática}
\usage{
AUTONUMBER(n)
}
\arguments{
\item{n}{Número de IDs a gerar.}
}
\value{
Vetor inteiro de 1 até \code{n}.
}
\description{
Gera sequência numérica automática
}
