% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{CHECK_IF_THEN}
\alias{CHECK_IF_THEN}
\title{Lógica condicional IF/THEN/ELSE}
\usage{
CHECK_IF_THEN(value)
}
\arguments{
\item{value}{Valor numérico a ser avaliado.}
}
\value{
\code{"High"} se \code{value > 10}, caso contrário \code{"Low"}.
}
\description{
Lógica condicional IF/THEN/ELSE
}
