% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{COMMENT_LEGAL}
\alias{COMMENT_LEGAL}
\title{Valida se um valor está entre os valores permitidos}
\usage{
COMMENT_LEGAL(value, legal_values)
}
\arguments{
\item{value}{Valor a ser verificado.}

\item{legal_values}{Vetor de valores válidos.}
}
\value{
Lógico: \code{TRUE} se o valor for permitido.
}
\description{
Valida se um valor está entre os valores permitidos
}
