% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{CREATE_FORM}
\alias{CREATE_FORM}
\title{Cria formulário de entrada de dados vazio}
\usage{
CREATE_FORM()
}
\value{
Data frame vazio com colunas padrão: ID, Name, Age, Gender, Date.
}
\description{
Cria formulário de entrada de dados vazio
}
