% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{CROSS_TAB}
\alias{CROSS_TAB}
\title{Tabulação cruzada com teste qui-quadrado}
\usage{
CROSS_TAB(data, var1, var2)
}
\arguments{
\item{data}{Data frame.}

\item{var1}{Nome da primeira variável (string).}

\item{var2}{Nome da segunda variável (string).}
}
\value{
Lista com \code{table} (tabulação cruzada) e \code{chi_square} (resultado do
teste qui-quadrado).
}
\description{
Tabulação cruzada com teste qui-quadrado
}
