% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episusTools-package.R
\docType{package}
\name{episusTools-package}
\alias{episusTools}
\alias{episusTools-package}
\title{episusTools: Ferramentas para Análise Epidemiológica}
\description{
Funções utilitárias para análise de dados epidemiológicos, desenvolvidas
para os treinandos do EpiSUS e outras áreas de análise de dados na SVSA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mariojrepisus/episusTools}
  \item Report bugs at \url{https://github.com/mariojrepisus/episusTools/issues}
}

}
\author{
\strong{Maintainer}: Mário Lima Júnior \email{mario.junior@saude.gov.br}

}
\keyword{internal}
