% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_est_ic95}
\alias{format_est_ic95}
\title{Formata estimativa com intervalo de confiança 95\%}
\usage{
format_est_ic95(est, ici, ics)
}
\arguments{
\item{est}{Estimativa central.}

\item{ici}{Limite inferior do IC 95\%.}

\item{ics}{Limite superior do IC 95\%.}
}
\value{
String no estilo \code{"est (ici - ics)"}.
}
\description{
Formata estimativa com intervalo de confiança 95\%
}
\examples{
format_est_ic95(1.5, 1.1, 2.1)
}
