% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_estimate}
\alias{format_estimate}
\title{Formata estimativas com 2 casas decimais}
\usage{
format_estimate(n)
}
\arguments{
\item{n}{Número a ser formatado.}
}
\value{
String formatada com 2 casas decimais.
}
\description{
Formata estimativas com 2 casas decimais
}
\examples{
format_estimate(1.2345)
}
