% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_n_int}
\alias{format_n_int}
\title{Formata números inteiros com separador de milhar}
\usage{
format_n_int(n)
}
\arguments{
\item{n}{Número a ser formatado.}
}
\value{
String formatada com separador de milhar e sem casas decimais.
}
\description{
Formata números inteiros com separador de milhar
}
\examples{
format_n_int(1234567)
}
