% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_n_perc}
\alias{format_n_perc}
\title{Formata número com percentual entre parênteses}
\usage{
format_n_perc(n, perc)
}
\arguments{
\item{n}{Número absoluto.}

\item{perc}{Percentual correspondente.}
}
\value{
String no estilo \code{"n (x\%)"}.
}
\description{
Formata número com percentual entre parênteses
}
\examples{
format_n_perc(250, 45.6)
}
