% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_perc}
\alias{format_perc}
\title{Formata percentual com 1 casa decimal}
\usage{
format_perc(n)
}
\arguments{
\item{n}{Número percentual a ser formatado.}
}
\value{
String formatada com 1 casa decimal.
}
\description{
Formata percentual com 1 casa decimal
}
\examples{
format_perc(45.678)
}
