% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_pvalor}
\alias{format_pvalor}
\title{Formata p-valor com 3 casas decimais}
\usage{
format_pvalor(n)
}
\arguments{
\item{n}{Número (p-valor) a ser formatado.}
}
\value{
String formatada com 3 casas decimais.
}
\description{
Formata p-valor com 3 casas decimais
}
\examples{
format_pvalor(0.0456)
}
