% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{FREQ}
\alias{FREQ}
\title{Tabela de frequências de uma variável}
\usage{
FREQ(data, variable)
}
\arguments{
\item{data}{Data frame.}

\item{variable}{Nome da coluna (string).}
}
\value{
Objeto \code{table} com as frequências.
}
\description{
Tabela de frequências de uma variável
}
