% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{IS_REQUIRED}
\alias{IS_REQUIRED}
\title{Verifica se um valor não é NA}
\usage{
IS_REQUIRED(value)
}
\arguments{
\item{value}{Valor a ser verificado.}
}
\value{
Lógico: \code{TRUE} se não for \code{NA}.
}
\description{
Verifica se um valor não é NA
}
