% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{LIST}
\alias{LIST}
\title{Lista os primeiros registros do data frame}
\usage{
LIST(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
As primeiras linhas do data frame.
}
\description{
Lista os primeiros registros do data frame
}
