% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{PLOT}
\alias{PLOT}
\title{Gráfico de barras de uma variável}
\usage{
PLOT(data, variable)
}
\arguments{
\item{data}{Data frame.}

\item{variable}{Nome da coluna (string).}
}
\value{
Objeto ggplot2.
}
\description{
Gráfico de barras de uma variável
}
