% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{SELECT}
\alias{SELECT}
\title{Filtra registros por uma condição}
\usage{
SELECT(data, condition)
}
\arguments{
\item{data}{Data frame.}

\item{condition}{String com a expressão de filtro (ex.: \code{"age > 18"}).}
}
\value{
Data frame filtrado.
}
\description{
Filtra registros por uma condição
}
