% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{SORT}
\alias{SORT}
\title{Ordena o data frame por uma variável}
\usage{
SORT(data, variable)
}
\arguments{
\item{data}{Data frame.}

\item{variable}{Nome da coluna para ordenar (string).}
}
\value{
Data frame ordenado.
}
\description{
Ordena o data frame por uma variável
}
