% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{SUMMARY}
\alias{SUMMARY}
\title{Resumo de todas as variáveis do data frame}
\usage{
SUMMARY(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Resumo estatístico via \code{summary()}.
}
\description{
Resumo de todas as variáveis do data frame
}
