% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_periodo_anos.R
\name{calcula_periodo_anos}
\alias{calcula_periodo_anos}
\title{Calcula o Período de Anos em um Data Frame}
\usage{
calcula_periodo_anos(banco, ano_col)
}
\arguments{
\item{banco}{Data frame que contém a coluna de anos.}

\item{ano_col}{Nome da coluna que contém os anos.}
}
\value{
Caractere com o período calculado.
}
\description{
Dado um data frame e o nome da coluna que contém os anos,
esta função retorna um texto representando o período (ex.: "2010-2023"
ou "2015" se houver apenas um ano).
}
\examples{
\dontrun{
  periodo <- calcula_periodo_anos(banco, ano_col = "ano")
  print(periodo)
}
}
