\name{CROSS_TAB}
\alias{CROSS_TAB}
\title{Tabulação cruzada com teste qui-quadrado}
\usage{CROSS_TAB(data, var1, var2)}
\arguments{
  \item{data}{Data frame.}
  \item{var1}{Nome da primeira variável (string).}
  \item{var2}{Nome da segunda variável (string).}
}
\value{Lista com `table` (tabulação cruzada) e `chi_square` (resultado do   teste qui-quadrado).}
