\name{episusTools-package}
\alias{episusTools-package}
\docType{package}
\title{episusTools: Ferramentas para Análise Epidemiológica}
\description{Funções utilitárias para análise de dados epidemiológicos, desenvolvidas para os treinandos do EpiSUS e outras áreas de análise de dados na SVSA.}
\keyword{internal}
