\name{format_est_ic95}
\alias{format_est_ic95}
\title{Formata estimativa com intervalo de confiança 95%}
\usage{format_est_ic95(est, ici, ics)}
\arguments{
  \item{est}{Estimativa central}
  \item{ici}{Limite inferior}
  \item{ics}{Limite superior}
}
\value{Texto formatado "est (ici - ics)"}
