\name{format_est_ic95}
\alias{format_est_ic95}
\title{Função format_est_ic95}
\description{}
\usage{format_est_ic95format_est_ic95((est, ici, ics){}
\arguments{
  \item{est}{Estimativa central}
  \item{ici}{Limite inferior}
  \item{ics}{Limite superior}
}
\value{Texto formatado "est (ici - ics)"}
