\name{format_estimate}
\alias{format_estimate}
\title{Formata estimativas com 2 casas decimais}
\description{ }
\usage{format_estimateformat_estimate((n) {}
\arguments{
  \item{n}{Número a ser formatado.}
}
\value{String formatada com 2 casas decimais.}
\examples{

format_estimate(1.2345)
}
