% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomp_sist_mult_dt.R
\name{incomp_sist_multi_dt}
\alias{incomp_sist_multi_dt}
\title{Avaliação de Completude de Variáveis de Data}
\usage{
incomp_sist_multi_dt(banco, variavel, nulo = NA, nivel_validos = 90)
}
\arguments{
\item{banco}{Data frame com os dados.}

\item{variavel}{Vetor de nomes das variáveis de data.}

\item{nulo}{Valor considerado nulo (padrão: \code{NA}).}

\item{nivel_validos}{Percentual mínimo de observações válidas para ser considerado adequado (padrão 90).}
}
\value{
Um \code{tibble} com a avaliação da completude para cada variável de data.
}
\description{
Esta função avalia a completude de variáveis de data (aquelas que iniciam com "DT")
em um data frame, transformando as datas e gerando um resumo.
}
\examples{
\dontrun{
  resultados <- incomp_sist_multi_dt(banco, c("DT_DATA1", "DT_DATA2"))
}
}
