% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{MATCH}
\alias{MATCH}
\title{Análise de tabela 2x2 para caso-controle}
\usage{
MATCH(case, control)
}
\arguments{
\item{case}{Contagem de casos expostos e não expostos.}

\item{control}{Contagem de controles expostos e não expostos.}
}
\value{
Resultado de \code{epiR::epi.2by2()}.
}
\description{
Wrapper de \code{epiR::epi.2by2()} configurado para estudo de caso-controle.
}
