\name{pacotes_uteis}
\alias{pacotes_uteis}
\title{Instala e Carrega Pacotes Úteis para Análise de Dados}
\description{ Instala o pacote `pacman` e o usa para carregar uma seleção de pacotes úteis para análise de dados em epidemiologia. Também instala pacotes específicos do Ministério da Saúde via GitHub.  }
\usage{pacotes_uteispacotes_uteis(() {}
\value{Invisível. Executada pelos efeitos colaterais de instalação e   carregamento de pacotes.}
\examples{

\dontrun{
  pacotes_uteis()
}
}
