% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacotes_uteis.R
\name{pacotes_uteis}
\alias{pacotes_uteis}
\title{Instala e Carrega Pacotes Úteis}
\usage{
pacotes_uteis()
}
\description{
Esta função instala o pacote \code{pacman} e, em seguida, utiliza-o para
carregar diversos pacotes úteis para análise de dados. Também instala
pacotes específicos do Ministério da Saúde via GitHub.
}
\examples{
\dontrun{
  pacotes_uteis()
}
}
