% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_info_classic_functions.R
\name{TABLE}
\alias{TABLE}
\title{Tabela 2x2 com OR/RR via epiR}
\usage{
TABLE(a, b, c, d)
}
\arguments{
\item{a}{Célula (exposto, doente).}

\item{b}{Célula (exposto, não doente).}

\item{c}{Célula (não exposto, doente).}

\item{d}{Célula (não exposto, não doente).}
}
\value{
Resultado de \code{epiR::epi.2by2()}.
}
\description{
Tabela 2x2 com OR/RR via epiR
}
