% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_episus.R
\name{tema_episus}
\alias{tema_episus}
\title{Tema Episus para ggplot2}
\usage{
tema_episus()
}
\value{
Um objeto de tema do ggplot2.
}
\description{
Retorna um tema customizado baseado no \code{theme_minimal} para uso no
\code{ggplot2}, conferindo um visual limpo e estilizado aos gráficos.
}
\examples{
ggplot(data) + geom_point(aes(x, y)) + tema_episus()
}
